refactor: asn 엔드포인트를 vendor/admin으로 변경한다.
